Resolve stack/heap crash in menu system
Changed the menu system to cache only the icons that are visible
on screen, instead of every icon for every menu entry. This relaxes
the limit on how many menu items there can be, so specifically the
custom bling menu can be much longer.

Also installed the mechanism to control the sizes of the stack and
the heap, in config/startup_config.h. However, the sizes currently
listed are the same as the old defaults, 0x2000 (8k) bytes for each.