Put the JoCo data we need in the BLE advertising data
The 'level' field was repurposed to lld (last_level_dispensed)
The 8 bit avatar field was replaces with a 16 bit score field.
Data offsets were adjusted accordingly in the data.
The new fields are read into the badge_db information for badges that are 'heard'

This has not yet been tested to make sure it's sent and received correctly.