Merge trinket dispensing flag with score in BLE advertising
It turns out that there was not enough room in the BLE data to add one more byte of data. By doing this, if the name is a full 8 bytes long, then it gets translated into a short naem format, and we don't watn that.

The solution is to simply adevrtise a flag that incicates that the badge has some outstanding 'credits' for trinkets which can be dispensed. This flag is the high but of the 16 bit unsigned score value. This works because the max score can never be more than 32767.