Unravel crypto functions from mbp_state
Created util/util_crypto to handle all the encryption and decryption
functions in a way that's compatible with the old crypto code that
was mixed in with state maintenance functions in mbp_state.

Revised mbp_state to use the new abstracted crypto functions.

Files are compatible in both directions, but the crypto functionality
is different in how it maintains the counter variable. The old code
went to some trouble to preserve the counter across multiple encrypts
of the data structure. However, it also generated a whole new nonce
for every encrypt, so there was no reason to persist the counter.
The new code starts the counter at zero (and generates a new nonce)
with every encrypt.

Usage instructions for util_crypto are in the .h file. It's expected
to be useful for JoCo database files and for secure transactions
over BLE.