Crypto state and score cleanup
Added macros in util_crypto.h to encapsulate some of the procedure for
accessing the util_crypto routines. Changed mbp_state.c and score_ble.c
to use them.

Switched from separate score_set and lld_set functions to a single
score_update function that does its own fetching of the data from the
mbp_state database. Changed mbp_state.c to use it. The new update
function rewrites the entire characteristic each time, and that's
the only place where the characteristic gets written.

Added a 4-byte secret cookie to the score read characteristic. This
doesn't really make it any harder to forge a response, but it does
make it easier for the reader to detect whether the response was
decrypted successfully.

Also added the device_id to the score read characteristic, just to
make sure we know who we're talking to.

Forced some of the structs to be packed, so the memory layout is
more deterministic (and compact).

Fixed the length fields in the characteristic setup.

Removed the spare field from the score read characteristic.